Return immediately on processFile decode errors instead of falling through
Each of the three decode steps (unmarshal storedEvent, look up the event
type, unmarshal its payload) reported a failure via sendErr but kept
executing on the zero-value/nil result, which could reach the handler with
a broken Event and then delete the file via the success path even though
it was never actually decoded — losing the event with no way to retry it,
contrary to processFile's own contract of leaving failed files in place.

Also fold eventbus_delivery_test.go into eventbus_test.go.
